fix effects to apply theme and write more documentation
