+++ docs/pages/installation/_meta.json
@@ -1,6 +1,16 @@
 {
 	"getting-started": "Getting Started",
-	"docker": "Docker",
-	"executable": "Executable",
+	"docker": {
+		"title": "Docker",
+		"theme": {
+			"toc": false
+		}
+	},
+	"executable": {
+		"title": "Executable",
+		"theme": {
+			"toc": false
+		}
+	},
 	"source": "Source"
 }
